Fixed-Width Type Consistency
Make sure some of the declarations + definitions are consistent, plus some other misc fixes

+++ include/JSystem/J3DGraphLoader/J3DMaterialFactory_v21.h
@@ -64,7 +64,7 @@ public:
     u16 countUniqueMaterials();
     u32 countTexGens(int) const;
     u32 countStages(int) const;
-    J3DMaterial* create(J3DMaterial*, int, unsigned long) const;
+    J3DMaterial* create(J3DMaterial*, int, u32) const;
     J3DGXColor newMatColor(int, int) const;
     u8 newColorChanNum(int) const;
     J3DColorChan newColorChan(int, int) const;
