+++ src/d/actor/d_a_tag_hint.cpp
@@ -478,7 +478,7 @@ void daTag_Hint_c::startProc() {
 }
 
 /* 000018A4-000018AC       .text next_msgStatus__12daTag_Hint_cFPUl */
-u16 daTag_Hint_c::next_msgStatus(unsigned long*) {
+u16 daTag_Hint_c::next_msgStatus(u32*) {
     return 16;
 }
