+++ src/d/actor/d_a_tag_island.cpp
@@ -136,7 +136,7 @@ void daTag_Island_c::demoInitProc() {
 }
 
 /* 00000550-00000558       .text next_msgStatus__14daTag_Island_cFPUl */
-int daTag_Island_c::next_msgStatus(unsigned long*) {
+int daTag_Island_c::next_msgStatus(u32*) {
     return fopMsgStts_MSG_ENDS_e;
 }
